chore(example-with-next): use app directory
