--- examples/next/src/pages/_app.tsx
@@ -1,9 +1,0 @@
-import { AppProps } from "next/app";
-
-import "../styles/index.css";
-
-const App = ({ Component, pageProps }: AppProps) => (
-  <Component {...pageProps} />
-);
-
-export default App;
